backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pillow==10.4.0
numpy==2.0.1
opencv-python-headless==4.10.0.84
